Adding new Featured functionality into plugin
Individual portfolios and portfolio terms (categories / tags) can now be featured. Featuring a portfolio makes them twice the width when viewed in a listing. This is a great way to showcase a single portfolio. The masonry / isotope behind the plugin that handles the dynamic filtering / layout has been refined also, it now works based on a spacer element so that the grid works fluidly across the site. also updated were some additional styles to make the portfolio lisiting better