Fixed bug in Content Type master class
Fixed a big that was affecting the filterable arguments that were passed to the master class. Meta box, meta field, post type, taxonomy and other settings can be filtered in the content type so that plugins can be extended. The bug was causing duplication of settings because the apply_filters function was being applied too many times (duplicating the additional settings each time it was called). There is only one place where the filters are called now and that is inside the filter_default_class_arguments method.